requirements.txt
PyQt6==6.8.0
pyusb==1.3.0
qasync==0.27.1
